+++ src/components/Button/Button.types.ts
@@ -1,0 +1,13 @@
+export enum ButtonType {
+  PRIMARY = 'primary',
+  DANGER = 'danger',
+  DEFAULT = 'default',
+  ACCENT = 'accent',
+}
+
+export interface ButtonProps {
+  type?: ButtonType
+  text: string
+  disabled?: boolean
+  onClick: () => void
+}